+++ api/routes/rbac/ModuleController.js
@@ -1,10 +1,12 @@
 import express from "express";
 import ModuleService from '@/services/rbac/ModuleService.js';
 import HttpSuccess from "@/utils/HttpSuccess.ts";
+import {checkPermission} from "@/middleware/permissionMiddleware.js";
 
 const router = express.Router();
 
 router.get("/",
+  checkPermission("module.read"),
   async (req, res) => {
     let modules = [];
     modules = await ModuleService.getModules();
@@ -16,6 +18,7 @@ router.get("/",
 
 
 router.post("/",
+  checkPermission("module.create"),
   async (req, res) => {
     const data = req.body;
     await ModuleService.saveModule(data);
@@ -27,6 +30,7 @@ router.post("/",
 
 router.patch(
   "/:id",
+  checkPermission("module.update"),
   async (req, res) => {
     const data = req.body;
     const { id } = req.params;
@@ -40,6 +44,7 @@ router.patch(
 
 router.delete(
   "/:id",
+  checkPermission("module.delete"),
   async (req, res) => {
     const {id} = req.params;
     await ModuleService.deleteModule(id);
